Extension: extract local LLM summarizer (oh-tab-b7qq)
